software/requirements.txt
pyecharts>=2.0.9
pyserial>=3.5
Requests>=2.32.5
